cpp4r: A C++ Interface for R's C Interface

Provides a header only, C++ interface to R's C interface. Compared to other approaches 'cpp4r' strives to be safe against long jumps from the C API as well as C++ exceptions, conform to normal R function semantics and supports interaction with 'ALTREP' vectors.

See also

Author

Maintainer: Mauricio Vargas Sepulveda m.vargas.sepulveda@gmail.com (ORCID)

Authors:

Posit Software, PBC (Original cpp11 package)