Vendor the cpp4r headers

Vendoring is the act of making your own copy of the 3rd party packages your project is using. It is often used in the go language community.

Usage

vendor(path = "./src/vendor")

Arguments

path

The directory to vendor the headers into

Value

The path to the vendored code (invisibly).

Details

This function vendors cpp4r into your package by copying the cpp4r headers into the inst/include folder of your package and adding 'cpp4r version: XYZ' to the top of the files, where XYZ is the version of cpp4r currently installed on your machine.

Note: vendoring places the responsibility of updating the code on you. Bugfixes and new features in cpp4r will not be available for your code until you run cpp_vendor() again.

Examples

# create a new directory
dir <- paste0(tempdir(), "/", gsub("\\s+|[[:punct:]]", "", Sys.time()))
dir.create(dir, recursive = TRUE)

# vendor the cpp4r headers into the directory
vendor(dir)

list.files(file.path(dir, "src", "vendor"))
#> character(0)

# cleanup
unlink(dir, recursive = TRUE)